// frontend/src/app/core/api/model/timetable-slot-detail-response.ts
/**
 * Training school system Backend API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { TimeSlotInfo } from './time-slot-info';


export interface TimetableSlotDetailResponse { 
    id?: number;
    subjectName?: string;
    teacherName?: string;
    roomName?: string;
    dayOfWeek?: string;
    periodicity?: TimetableSlotDetailResponse.PeriodicityEnum;
    timeSlot?: TimeSlotInfo;
}
export namespace TimetableSlotDetailResponse {
    export const PeriodicityEnum = {
        Normal: 'NORMAL',
        WeekA: 'WEEK_A',
        WeekB: 'WEEK_B'
    } as const;
    export type PeriodicityEnum = typeof PeriodicityEnum[keyof typeof PeriodicityEnum];
}